+++ Changelog.md
@@ -1,6 +1,9 @@
 ## Ignition Transport 7
 
-### Ignition Transport 7.X.X
+### Ignition Transport 7.1.0
+
+1. Added method for determining if a throttled publisher is ready to publish.
+    * [Pull request 395](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/395)
 
 ### Ignition Transport 7.0.0